Fix GGUF timestamp fingerprint on macOS
Cast the filesystem clock count to an explicit integer type before passing it to std::to_string, avoiding an ambiguous libc++ overload on Xcode 15.4.

+++ src/framework/assets/tensor_source.cpp
@@ -1217,7 +1217,8 @@ std::filesystem::path materialize_gguf_sidecars(const std::filesystem::path & pa
     }
     const auto fingerprint_source = canonical.string() + ":" +
         std::to_string(std::filesystem::file_size(canonical)) + ":" +
-        std::to_string(std::filesystem::last_write_time(canonical).time_since_epoch().count());
+        std::to_string(static_cast<long long>(
+            std::filesystem::last_write_time(canonical).time_since_epoch().count()));
     std::ostringstream fingerprint;
     fingerprint << std::hex << std::hash<std::string>{}(fingerprint_source);
     const auto root = std::filesystem::temp_directory_path() / "audiocpp-gguf" / fingerprint.str();
